+++ src/main/java/info/faceland/strife/data/LoadedStatBoost.java
@@ -11,7 +11,7 @@ public class LoadedStatBoost {
   private final String creator;
   private final int duration;
   private final int announceInterval;
-  private final Map<StrifeStat, Double> stats;
+  private final Map<StrifeStat, Float> stats;
   private final List<String> announceStart;
   private final List<String> announceRun;
   private final List<String> announceEnd;
@@ -26,7 +26,7 @@ public class LoadedStatBoost {
     this.announceEnd = new ArrayList<>();
   }
 
-  public Map<StrifeStat, Double> getStats() {
+  public Map<StrifeStat, Float> getStats() {
     return stats;
   }
