+++ src/main/java/info/faceland/strife/managers/StrifeMobManager.java
@@ -60,7 +60,7 @@ public class StrifeMobManager {
     return initialSize - trackedEntities.size();
   }
 
-  public void setEntityStats(LivingEntity entity, Map<StrifeStat, Double> statMap) {
+  public void setEntityStats(LivingEntity entity, Map<StrifeStat, Float> statMap) {
     StrifeMob strifeMob = getStatMob(entity);
     strifeMob.setStats(statMap);
     trackedEntities.put(entity.getUniqueId(), strifeMob);
